refactor: 토큰엔티티 추가 및 적용
