--- data/account/src/main/java/com/into/websoso/data/account/AccountEntity.kt
@@ -1,7 +1,0 @@
-package com.into.websoso.data.account
-
-data class AccountEntity(
-    val accessToken: String,
-    val refreshToken: String,
-    val isRegister: Boolean,
-)
