+++ data/account/src/main/java/com/into/websoso/data/account/model/AccountEntity.kt
@@ -1,0 +1,6 @@
+package com.into.websoso.data.account.model
+
+data class AccountEntity(
+    val token: TokenEntity,
+    val isRegister: Boolean,
+)
