+++ data/account/src/main/java/com/into/websoso/data/account/model/TokenEntity.kt
@@ -1,0 +1,6 @@
+package com.into.websoso.data.account.model
+
+data class TokenEntity(
+    val accessToken: String,
+    val refreshToken: String,
+)
